MemoryAccessor WriteWord added /unit test
memory size of byte made to return size_t

WriteBack in progress for CF

+++ NES/MOS6502/Memory.h
@@ -70,7 +70,7 @@ namespace mos6502
 			return addressSize;
 		}
 
-		inline unsigned int getTotalBytes() const
+		inline size_t getTotalBytes() const
 		{
 			return memory.size();
 		}
